+++ Transform/PM/pm.cpp
@@ -239,18 +239,18 @@ void simuSfence() {
   }
 }
 
-void simuFlushOpt(void* ptr, size_t len) {
+void simuFlushOpt(void* ptr, int len) {
   if (my_flag) {
     pthread_mutex_lock(&mtx);
-    my_printf("Simulated F-L-U-S-H: %p LEN: %zu; ", ptr, len);
+    my_printf("Simulated F-L-U-S-H: %p LEN: %d; ", ptr, len);
     pthread_mutex_unlock(&mtx);
   }
 }
 
-void simuFlush(void* ptr, size_t len) {
+void simuFlush(void* ptr, int len) {
   if (my_flag) {
       pthread_mutex_lock(&mtx);
-      my_printf("Simulated F-L-U-S-H: %p LEN: %zu; ", ptr, len);
+      my_printf("Simulated F-L-U-S-H: %p LEN: %d; ", ptr, len);
       pthread_mutex_unlock(&mtx);
   }
 }
